multitenancy fixes
