+++ src/SWP/Bundle/CoreBundle/Resources/config/doctrine/Rule.orm.yml
@@ -1,4 +1,4 @@
-SWP\Bundle\CoreBundle\Entity\Rule:
+SWP\Bundle\CoreBundle\Model\Rule:
     type: entity
     table: swp_rule
     fields:
