+++ src/SWP/Bundle/CoreBundle/Resources/config/serializer/Model.Rule.yml
@@ -1,4 +1,4 @@
-SWP\Bundle\CoreBundle\Entity\Rule:
+SWP\Bundle\CoreBundle\Model\Rule:
     exclusion_policy: ALL
     relations:
         - rel: self
